metadata delete method update

+++ backend/src/main/java/com/bulletjournal/templates/repository/ChoiceMetadataKeywordDaoJpa.java
@@ -1,5 +1,6 @@
 package com.bulletjournal.templates.repository;
 
+import com.bulletjournal.exceptions.BadRequestException;
 import com.bulletjournal.exceptions.ResourceNotFoundException;
 import com.bulletjournal.templates.repository.model.Choice;
 import com.bulletjournal.templates.repository.model.ChoiceMetadataKeyword;
@@ -9,6 +10,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Repository
 public class ChoiceMetadataKeywordDaoJpa {
@@ -47,7 +50,11 @@ public class ChoiceMetadataKeywordDaoJpa {
 
     @Transactional(rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
     public void deleteByKeywords(List<String> keywords) {
-        List<ChoiceMetadataKeyword> list = choiceMetadataKeywordRepository.findAllById(keywords);
+        if (keywords.isEmpty()) {
+            throw new BadRequestException("keywords is empty");
+        }
+        List<ChoiceMetadataKeyword> list = choiceMetadataKeywordRepository.findAllById(keywords)
+                .stream().filter(Objects::nonNull).collect(Collectors.toList());
         choiceMetadataKeywordRepository.deleteAll(list);
     }
 }
