+++ backend/src/main/java/com/bulletjournal/templates/repository/StepMetadataKeywordDaoJpa.java
@@ -1,5 +1,6 @@
 package com.bulletjournal.templates.repository;
 
+import com.bulletjournal.exceptions.BadRequestException;
 import com.bulletjournal.exceptions.ResourceNotFoundException;
 import com.bulletjournal.templates.repository.model.Step;
 import com.bulletjournal.templates.repository.model.StepMetadataKeyword;
@@ -9,6 +10,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Repository
 public class StepMetadataKeywordDaoJpa {
@@ -46,7 +49,11 @@ public class StepMetadataKeywordDaoJpa {
 
     @Transactional(rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
     public void deleteByKeywords(List<String> keywords) {
-        List<StepMetadataKeyword> list = stepMetadataKeywordRepository.findAllById(keywords);
+        if (keywords.isEmpty()) {
+            throw new BadRequestException("keywords is empty");
+        }
+        List<StepMetadataKeyword> list = stepMetadataKeywordRepository.findAllById(keywords)
+                .stream().filter(Objects::nonNull).collect(Collectors.toList());
         stepMetadataKeywordRepository.deleteAll(list);
     }
 }
